refactor : 역할이 중복되는 RacingGame, RacingService 를 RacingGame 로 통합
