+++ src/main/java/racingcar/util/MovingNumberPolicy.java
@@ -2,5 +2,5 @@ package racingcar.util;
 
 public interface MovingNumberPolicy {
 
-	int generate(int bound);
+	int generate();
 }
